+++ Include/Flux/UI/UserInterface.h
@@ -1,7 +1,6 @@
-﻿#pragma once
+#pragma once
 
 #include <Flux/UI/Component.h>
-#include <Flux/UI/MasterView.h>
 #include <Flux/UI/CursorManager.h>
 #include <Flux/UI/Stack.h>
 #include <Flux/UI/Color.h>
