+++ Source/Application/Window/Window.cpp
@@ -1,0 +1,12 @@
+#include <Application/Window/Window.h>
+#include <Application/Window/OpenGLWindow.h>
+#include <Application/Window/MetalWindow.h>
+#include <skia/include/core/SkFont.h>
+
+namespace Flux {
+    
+    Window::~Window() {
+        glfwDestroyWindow(handle);
+    }
+    
+}
